# Faster compilation of bevy_diagnostic