--- a/crates/bevy_diagnostic/src/diagnostic.rs
+++ b/crates/bevy_diagnostic/src/diagnostic.rs
@@ -1,5 +1,5 @@
-use bevy_utils::{Duration, HashMap, Instant, Uuid};
-use std::collections::{BTreeSet, VecDeque};
+use bevy_utils::{Duration, Instant, StableHashMap, Uuid};
+use std::collections::VecDeque;
 
 /// Unique identifier for a [Diagnostic]
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq, PartialOrd, Ord)]
@@ -101,13 +101,13 @@ impl Diagnostic {
 /// A collection of [Diagnostic]s
 #[derive(Debug, Default)]
 pub struct Diagnostics {
-    diagnostics: HashMap<DiagnosticId, Diagnostic>,
-    ordered_diagnostics: BTreeSet<DiagnosticId>,
+    // This uses a [`StableHashMap`] to ensure that the iteration order is deterministic between
+    // runs when all diagnostics are inserted in the same order.
+    diagnostics: StableHashMap<DiagnosticId, Diagnostic>,
 }
 
 impl Diagnostics {
     pub fn add(&mut self, diagnostic: Diagnostic) {
-        self.ordered_diagnostics.insert(diagnostic.id);
         self.diagnostics.insert(diagnostic.id, diagnostic);
     }
 
@@ -133,11 +133,5 @@ impl Diagnostics {
 
     pub fn iter(&self) -> impl Iterator<Item = &Diagnostic> {
         self.diagnostics.values()
-    }
-
-    pub fn ordered_iter(&self) -> impl Iterator<Item = &Diagnostic> {
-        self.ordered_diagnostics
-            .iter()
-            .filter_map(move |k| self.diagnostics.get(k))
     }
 }