--- a/crates/bevy_utils/src/lib.rs
+++ b/crates/bevy_utils/src/lib.rs
@@ -11,14 +11,45 @@ pub type BoxedFuture<'a, T> = Pin<Box<dyn Future<Output = T> + Send + 'a>>;
 #[cfg(target_arch = "wasm32")]
 pub type BoxedFuture<'a, T> = Pin<Box<dyn Future<Output = T> + 'a>>;
 
+/// A hasher builder that will create a fixed hasher.
+#[derive(Default)]
+pub struct FixedState;
+
+impl std::hash::BuildHasher for FixedState {
+    type Hasher = AHasher;
+
+    #[inline]
+    fn build_hasher(&self) -> AHasher {
+        AHasher::new_with_keys(0, 0)
+    }
+}
+
 /// A std hash map implementing AHash, a high speed keyed hashing algorithm
 /// intended for use in in-memory hashmaps.
 ///
 /// AHash is designed for performance and is NOT cryptographically secure.
 pub type HashMap<K, V> = std::collections::HashMap<K, V, RandomState>;
 
+/// A stable std hash map implementing AHash, a high speed keyed hashing algorithm
+/// intended for use in in-memory hashmaps.
+///
+/// Unlike [`HashMap`] this has an iteration order that only depends on the order
+/// of insertions and deletions and not a random source.
+///
+/// AHash is designed for performance and is NOT cryptographically secure.
+pub type StableHashMap<K, V> = std::collections::HashMap<K, V, FixedState>;
+
 /// A std hash set implementing AHash, a high speed keyed hashing algorithm
 /// intended for use in in-memory hashmaps.
 ///
 /// AHash is designed for performance and is NOT cryptographically secure.
 pub type HashSet<K> = std::collections::HashSet<K, RandomState>;
+
+/// A stable std hash set implementing AHash, a high speed keyed hashing algorithm
+/// intended for use in in-memory hashmaps.
+///
+/// Unlike [`HashSet`] this has an iteration order that only depends on the order
+/// of insertions and deletions and not a random source.
+///
+/// AHash is designed for performance and is NOT cryptographically secure.
+pub type StableHashSet<K> = std::collections::HashSet<K, FixedState>;